Removed unnecessary files and made improvements to UX
